Initial commit of fakeredis
This is a very rough initial commit.
It only supports the list commands, and omits
a lot of error conditions.  I eventually plan to
add all the types, and handle all the various error
conditions.

I've also added unittests that can optionally
run with the real redis module (and a real redis
server) to ensure parity between the real redis
client and fake redis.