Proper implementation of getbit/setbit
While getbit/setbit worked fine together,
they did not work well when using get().
This is because the implementation of bits
leaks into how get() is implemented.  The
basic problem was that I was treating setting
bits as a number when redis actually treats
the bits as char[].  Also the bit numbering
goes left to right so something like::

  setbit foo 3 1

actually means::

  00100000

and not::

  00000100

I was previously assuming that setting a bit
was basically (1 << offset), but that's a right
to left ordering.