Check keys before delete in  _update_expired_keys
I was running into an issue when use fakeredis and freezegun to manipulate dates in a unit test. The _StrKeyDict._update_expired_keys was trying to delete keys that appeared the _ex_keys dict, but somehow the keys did not appear in the underlying _dict of the class instance.
This just updates the method to verify that the key is contained in the _ex_keys and _dict objects before calling del.