Add type checks on ZSET operations
Any operation that queries or mutates a ZSET will raise an appropriate
redis.ResponseError if the key is already present with a different type.
New tests are included to test this. ZINTERSTORE and ZUNIONSTORE have
special treatment to allow SET types as well (undocumented feature of
redis).

The goal is to later extend this to other types as well.