Change pubsub matching to use _compile_pattern
This corrects the previous pattern -> regex generation, which was badly
broken (didn't escape regex special chars, didn't anchor at the end,
for some reason made [] groups optional matches etc).

In order to make this work, it was also necessary to redo some of the
pubsub code to treat channel names as bytes when decode_responses=False
and strings when it is True. That fixes a whole class of bugs where
channel names with non-UTF-8 in them would go wrong.