Handle decode_responses correctly for score_cast_func
When decode_responses is true, the score_cast_func is passed text, not
bytes. To handle this, the score_cast_func handling code was unified in
one helper function which deals with the different cases. It also has a
cut-through path for the default score_cast_func to avoid a lot of
unnecessarily float -> bytes -> text -> float casting.

The unit test is made stricted to check the type exactly.