Optimize `_scan()`.
In case that a regular expression is specified, it is not needed to copy all the keys if some of them do not match the regular expression.
If a regular expression is not provided, there is no need to iterate over the slice. Instead, we simply copy the entire slice.
This is useful in cases where we're dealing with a large number of keys.