Make the disconnected pubsub tests pass
Now additionally raises a connection error when disconnect and
- trying to perform the initial connection
- reading a response from an empty queue

The test also had to be updated because it didn't match behaviour with a
real redis server. One test had to be removed because there is no way to
publish to a disconnected server, and publishing to a connected server
immediately queues the result to all clients because it can be
disconnected.