Minor optimisations to zset
Instead of storing a separate SortedDict sorted lexically, it just
stores a regular dict to look up scores. The *BYLEX commands now take
advantage of the assumption that all scores are equal by searching in
the _byscore list, using the score of the first key.

This makes a very small improvement in performance, and also reduces
memory usage.