Support NX/XX/CH flags in ZADD command (#247)
Support NX/XX/CH flags in ZADD command

Add support for the new `NX`, `XX`, and `CH` flags in the `ZADD`
command, which can only be provided with `redis-py` version 3 installed.

This fixes a subset of what's filed in #232. `INCR` flag will be
implemented in a different pull request, as it's completely different to
handle.