Make unit test pass again
Rather than trying to restore the hackery that made DecodeMixin work
before by running all tests and then adjusting the assertions, we just
run a handful of tests to check that the decoding machinery is working.
This should still give good coverage now that fakeredis itself handles
the decoding in a simple generic way.