Add support for SCRIPT EXISTS subcommand
I've added support for the SCRIPT EXISTS subcommand in the SCRIPTS
command handler. The subcommand takes in a sequence of sha1 hash values
of Lua scripts and checks if scripts with those sha1 hash values are
currently registered with Redis.

The command will return a redis List with the integers 1 and 0 to
indicate which sha1 hash is known to Redis. In accordance with Redis,
this subcommand returns an empty List if no hash values were passed as
arguments to the subcommand.