Align `FakeConnection` constructor signature to base class.
As `FakeConnection` is a subclass of the `Connection` class in
`redis-py`, in order to honor OOP substitutability principle, the
signature of the constructor should be aligned with the one of the base
class. That means it should contain at least all the parameters of the
base class.

This automatically fixes a problem occurring when the connection object
is dynamically instantiated within the context of the default connection
pool class in `redis-py`, named `ConnectionPool`, and when, in turn, the
latter is dynamically instantiated using its standard `from_url` factory
method. In such scenario the instantiation of the connection object
fails because the `FakeConnection` class misses both `host` and `port`
parameters in its constructor signature.

That happens because the `from_url` factory method above, given a
canonical Redis URL, guesses several connection parameters values, among
which `host`, `username`, `password`, `path`, and `db`, which values
override any other corresponding value passed to the factory method
itself via the connection variadic arguments.

Aligning the `FakeConnection` constructor signature to the one of its
base class, by using variadic arguments, is helpful in contexts where
there's no direct control over the instantiantiation of Redis clients,
connections and connection pools objects (e.g. `django-redis` library):
whilst preserving current behavior, this avoids to workaround the issue
by defining either custom connection factories or custom connection
pools.